scanner-service/requirements.txt
pillow
numpy
paho-mqtt
